hubspot-sheets-sync/pi-deployment/requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
hubspot-api-client>=8.0.0
python-dotenv>=1.0.0
